+++ R/tna-package.R
@@ -4,8 +4,8 @@
 #' @family basic
 #' @description Provides tools for performing transition network analysis (TNA),
 #' including functions for building TNA models, plotting transition networks,
-#' and calculating centrality measures. The package relies on the `qgraph`
-#' and `igraph` for network plotting and centrality measure calculations.
+#' and calculating centrality measures. The package relies on the  `igraph` for 
+#' centrality measure calculations.
 #'
 #' @author Sonsoles López-Pernas, Santtu Tikka, Mohammed Saqr
 #' @references