+++ NEWS.md
@@ -4,6 +4,8 @@
 * Fixed an issue with `bootstrap()` when the model contained weights that were zero.
 * Fixed issues with several distance measures in `cluster_sequences()`.
 * Fixed edge ordering in `betweenness_network()`.
+* The packages no longer depends on `qgraph`.
+* Network plotting functionality now relies on `cograph` instead.
 
 # tna 1.2.0